moved block tracker to its own package
